Separated Admin Routes from regular routes
